+++ lib/eve_online.rb
@@ -48,6 +48,7 @@ require 'eve_online/esi/corporation'
 require 'eve_online/esi/corporation_industry_jobs'
 require 'eve_online/esi/corporation_blueprints'
 require 'eve_online/esi/corporation_orders'
+require 'eve_online/esi/corporation_killmails_recent'
 
 require 'eve_online/esi/alliances'
 require 'eve_online/esi/alliance'